Writes Metadata

Utilizes exiftool to write metadata to generated plots. Writes the R version, lattice version, latticeExtra version, BoutrosLab.plotting.general version, BoutrosLab.plotting.survival version, operating system, machine, author, image description.

Usage

write.metadata(
  filename = NULL,
  description = NULL,
  verbose = FALSE
  );

Arguments

filename

Filename for output, or if NULL (default value) returns image unchanged.

description

Short description of image; default NULL

verbose

Option to standard output; default FALSE

Value

If filename is NULL, returns the image unchanged. If description is NULL, then the image is returned without the description tag.

Note: an easy way to view the metadata is by using the exiftool command.

Author

Esther Jung